requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
jinja2
python-multipart
markdown
markupsafe
openai
weasyprint==61.2
